fix: typing Интерфейс вынес в отдельный файл и подключил к компонентам Исправил ошибки которые были указаны Благодарю!
