+++ src/app/interfaces/user-interface.ts
@@ -1,0 +1,26 @@
+export interface User {
+  id: number;
+  name: string;
+  username: string;
+  email: string;
+  address: {
+    street: string;
+    suite: string;
+    city: string;
+    zipcode: string;
+  }
+  phone: string;
+  website: string;
+  company: {
+    name: string;
+    catchPhrase: string;
+    bs: string;
+  }
+}
+
+export interface Todo {
+  userId: number,
+  id: number,
+  title: string,
+  completed: boolean
+}
