+++ src/app/todo-list/todo-card/todo-card.component.html
@@ -7,5 +7,5 @@
 
   <p>ЗАДАЧА ЗАВЕРШЕНА:</p>
   <div>{{todo!.completed ? 'ДА' : 'НЕТ'}}</div>
-  <button (click)=onDeleteTodo(todo!.id)>УДАЛИТЬ ЗАДАЧУ</button>
+  <button (click)=onDeleteTodo(todo.id)>УДАЛИТЬ ЗАДАЧУ</button>
 </div>
