+++ src/app/users-list/user-card/user-card.component.html
@@ -3,5 +3,5 @@
   <p>{{$any(user).email}}</p>
   <p>{{$any(user).website}}</p>
   <p>{{$any(user).company.name}}</p>
-  <button class="click" (click)=onDeleteUser(user!.id)>УДАЛИТЬ ЮЗЕРА</button>
+  <button class="click" (click)=onDeleteUser(user.id)>УДАЛИТЬ ЮЗЕРА</button>
 </div>
