+++ src/app/users-list/users-list.component.ts
@@ -2,33 +2,14 @@ import {Component, inject} from "@angular/core";
 import {NgFor} from "@angular/common";
 import {UsersApiService} from "../users-api.service";
 import {UserCardComponent} from "./user-card/user-card.component";
-
-export interface User {
-    id: number;
-    name: string;
-    username: string;
-    email: string;
-    address: {
-        street: string;
-        suite: string;
-        city: string;
-        zipcode: string;
-        }
-    phone: string;
-    website: string;
-    company: {
-        name: string;
-        catchPhrase: string;
-        bs: string;
-    }
-}
+import {User} from "../interfaces/user-interface";
 
 @Component({
   selector: 'app-users-list',
   templateUrl: './users-list.component.html',
   standalone: true,
   styleUrl: './users-list.component.scss',
-  imports: [NgFor, UserCardComponent],
+  imports: [NgFor, UserCardComponent]
 })
 
 export class UsersListComponent {
